+++ src/annbatch/__init__.py
@@ -5,7 +5,7 @@ from importlib.metadata import version
 from . import abc, types
 from .io import DatasetCollection, write_sharded
 from .loader import Loader
-from .samplers._chunk_sampler import ChunkSampler, ChunkSamplerDistributed
+from .samplers import ChunkSampler, ChunkSamplerDistributed, ChunkSamplerWithReplacement
 
 __version__ = version("annbatch")
 
@@ -16,5 +16,6 @@ __all__ = [
     "write_sharded",
     "ChunkSampler",
     "ChunkSamplerDistributed",
+    "ChunkSamplerWithReplacement",
     "abc",
 ]