+++ src/annbatch/samplers/__init__.py
@@ -1,6 +1,8 @@
-from ._chunk_sampler import ChunkSampler, ChunkSamplerDistributed
+from ._chunk_sampler import ChunkSampler, ChunkSamplerWithReplacement
+from ._chunk_sampler_distributed import ChunkSamplerDistributed
 
 __all__ = [
     "ChunkSampler",
     "ChunkSamplerDistributed",
+    "ChunkSamplerWithReplacement",
 ]